requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==11.2.1
environs[django]==9.3.2
django-phonenumber-field==6.4.0
marshmallow==3.0.0
phonenumbers==9.0.6
djangorestframework==3.13.0
requests==2.31.0
geopy==2.4.1
setuptools==80.9.0